Initial debianization of istio-cni
This debian directory is added after importing
the istio source release tarball from github,
and issuing `gbp import-orig --pristine-tar`
to import it to the pristine-tar and upstream
branches.

The idea is to have a generic istio repository
that builds binaries, using control to create
specialized packages. The first package that
we need is the istio-cni one, that should simply
ship two binaries to kubernetes nodes
(istio-cni and istio-iptables).

The debian repack import workflow has not been used
since the Istio Makefile config is already downloading
various bits from the Google API / Cloud, and to
avoid interfering with it as much as possible
(the vendor deps are very few as far as I can see).
If needed the repack workflow can be added, I'll
go with any decision we'll reach discussing this patch.

Bug: T297612
Change-Id: Iab50b8b3515f18ec597fc03b18dc392d894c927e